tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
pydantic==1.8.2
elasticsearch~=7.15.1
python-dotenv==0.19.2
aioredis==1.3.1
redis==4.0.2
requests==2.26.0